// configurator_frontend/src/payload-types.ts
export interface Media {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    admin_thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface ManualSelectAttribute {
  name: string;
  code: string;
  options: (TextOption | ColorOption)[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualSelectAttribute';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TextOption".
 */
export interface TextOption {
  name: string;
  default?: boolean | null;
  image?: string | Media | null;
  code: string;
  actions: (
    | ManualShowAction
    | ManualHideAction
    | ManualSetColorAction
    | SetUserImageAction
    | SetTextureAction
    | SetColorAction
    | ShowPartAction
    | ManualSetTextureAction
    | UnsetAttributeAction
  )[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'text';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualShowAction".
 */
export interface ManualShowAction {
  nodes?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualShow';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualHideAction".
 */
export interface ManualHideAction {
  nodes?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualHide';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualSetColorAction".
 */
export interface ManualSetColorAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  colorCode: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualSetColor';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetUserImageAction".
 */
export interface SetUserImageAction {
  width: number;
  height: number;
  material: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setUserImage';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetTextureAction".
 */
export interface SetTextureAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  wrapS: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  wrapT: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  repeatX: number;
  repeatY: number;
  offsetX: number;
  offsetY: number;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setTexture';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetColorAction".
 */
export interface SetColorAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setColorAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ShowPartAction".
 */
export interface ShowPartAction {
  id?: string | null;
  blockName?: string | null;
  blockType: 'showPart';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualSetTextureAction".
 */
export interface ManualSetTextureAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  colorMap: string | Media;
  color?: string | null;
  aoMap?: string | Media | null;
  aoIntensity?: number | null;
  displacementMap?: string | Media | null;
  displacementScale?: number | null;
  normalMap?: string | Media | null;
  normalScaleX?: number | null;
  normalScaleY?: number | null;
  roughnessMap?: string | Media | null;
  roughness?: number | null;
  metalnessMap?: string | Media | null;
  metalness?: number | null;
  bumpMap?: string | Media | null;
  bumpScale?: number | null;
  emissiveMap?: string | Media | null;
  emissiveColor?: string | null;
  emissiveIntensity?: number | null;
  wrapS: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  wrapT: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  repeatX: number;
  repeatY: number;
  offsetX: number;
  offsetY: number;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualSetTextureAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "UnsetAttributeAction".
 */
export interface UnsetAttributeAction {
  attributeCode: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'unsetAttributeAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ColorOption".
 */
export interface ColorOption {
  name: string;
  default?: boolean | null;
  code: string;
  colorCode: string;
  actions: (
    | ManualShowAction
    | ManualHideAction
    | ManualSetColorAction
    | SetUserImageAction
    | SetTextureAction
    | SetColorAction
    | ShowPartAction
    | ManualSetTextureAction
    | UnsetAttributeAction
  )[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'color';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "UserImageAttribute".
 */
export interface UserImageAttribute {
  name: string;
  code: string;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'userImage';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TextureAttribute".
 */
export interface TextureAttribute {
  name: string;
  code: string;
  tags?: (string | Tag)[] | null;
  defaultValue?: (string | null) | Texture;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'textureAttribute';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tags".
 */
export interface Tag {
  id: string;
  tag: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "textures".
 */
export interface Texture {
  id: string;
  title: string;
  code: string;
  image: string | Media;
  tags?: (string | Tag)[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ColorAttribute".
 */
export interface ColorAttribute {
  name: string;
  code: string;
  colors?: (string | Color)[] | null;
  customColors?: boolean | null;
  defaultValue?: (string | null) | Color;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'colorAttribute';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "colors".
 */
export interface Color {
  id: string;
  name: string;
  colorCode: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "PartsAttribute".
 */
export interface PartsAttribute {
  name: string;
  code: string;
  parts?:
    | {
        name: string;
        code: string;
        image?: string | Media | null;
        nodes?:
          | {
              name: string;
              id?: string | null;
            }[]
          | null;
        default: boolean;
        id?: string | null;
      }[]
    | null;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'partsAttribute';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BuildingImageAttribute".
 */
export interface BuildingImageAttribute {
  name: string;
  code: string;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'buildingImage';
}